# ironbridge-rust/docs/assets/openapi.yaml
openapi: 3.0.3
info:
  title: IronBridge API
  description: |
    REST API for IronBridge (Chat Session Manager) - a unified platform for harvesting, 
    managing, and analyzing AI chat sessions across multiple providers and tools.

    ## Authentication

    Currently the API supports local access without authentication. Enterprise features
    include SSO/SAML authentication and API key-based access.

    ## Rate Limiting

    Rate limiting is optional and configurable per-installation.

    ## WebSocket

    Real-time updates are available via WebSocket at `ws://<host>/ws`. Note that
    `/ws` is mounted at the server root, not under the `/api` base URL below.

    Client frames are tagged JSON: `subscribe`, `unsubscribe`, `ping`,
    `sync_request`, `stream_start`, `stream_cancel`, `stream_input` and
    `agent_command`. Every one of them is answered.

    `sync_request` returns the same delta as `GET /sync/delta?from=N`, as one
    `sync_event` per change in version order. Two cases answer with an error
    instead of an empty result, because sending nothing would read as "you are
    up to date": `history_truncated` when the requested version has aged out of
    the retained event history, and `version_ahead` when the client's version
    exceeds the server's. Both mean the client needs a full snapshot.

    The stream operations and `agent_command` are **not implemented** on this
    server and answer `stream_error` / `error` with code `unsupported`. They are
...

...

...
servers:
...
paths:
...
  /swarms:
    get:
      tags: [Swarms]
      summary: List swarms
      operationId: listSwarms
      responses:
        "200":
          description: Swarms
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
    post:
      tags: [Swarms]
      summary: Create a swarm
      operationId: createSwarm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, orchestration, agents]
              properties:
                name:
                  type: string
                description:
                  type: string
                orchestration:
                  type: string
                  enum: [sequential, parallel, hierarchical, debate]
                agents:
                  type: array
                  items:
                    type: object
                max_iterations:
                  type: integer
      responses:
        "200":
          description: Swarm created
...
  /sync/delta:
    servers: *rootServer
    get:
      tags: [Sync]
      summary: Changes since a version
      description: >
        Note the mixed casing -- `from_version` and `to_version` are
        snake_case where the rest of the payload is not. That is what the
        server sends.
      operationId: getSyncDelta
      parameters:
        - name: since
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Delta
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SyncDelta"
...
  /sync/event:
    servers: *rootServer
    post:
      tags: [Sync]
      summary: Record one sync event
      operationId: postSyncEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Event recorded
  /sync/batch:
    servers: *rootServer
    post:
      tags: [Sync]
      summary: Record several sync events
      operationId: postSyncBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Events recorded
...
components:
...
  schemas:
...
    ApiEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          description: The payload. Absent on failure.
        error:
          type: string
          nullable: true
...
    SyncDelta:
      type: object
      properties:
        from_version:
          type: integer
        to_version:
          type: integer
        created:
          type: array
          items:
            type: object
        updated:
          type: array
          items:
            type: object
        deleted:
          type: array
          items:
            type: object
        timestamp:
          type: integer
          format: int64
...